+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # 1.x branch
+## 1.6 branch
+### 1.6.0
+* removed the backward compatibility with the previous `Exceptionist::inArray()`
+    method, provided by `__call()`.
+
 ## 1.5 branch
 ### 1.5.8
 * `Entity::set()` can set null or empty values and no longer throws an exception.